refactor(website): extract gallery items into src/data/gallery.ts
Move the inline `GalleryItem` interface and items array out of
GallerySection.vue and into a dedicated data module. Adds a stable
`id` slug per item, an optional `visible` flag with a `visibleGalleryItems`
filtered export, and a `getGalleryItemById` lookup helper. Sibling
components now import the type from the data file instead of from a
.vue file.
